Update mynintendo.py
Woo! That took too much work to fix an issue that ended up with such a simple fix...story of coding.

Script now has access to the dynamically updated javascript instead of just static HTML, which caused it to stop working. Uses a chromedriver and everything...

Will update readme to account for this.